refactor(novel): prevent using route handler in server component
