+++ package.json
@@ -21,6 +21,7 @@
 		"@fortawesome/fontawesome-svg-core": "^6.5.2",
 		"@fortawesome/free-solid-svg-icons": "^6.5.2",
 		"@fortawesome/react-fontawesome": "^0.2.0",
+		"@neondatabase/serverless": "^0.9.3",
 		"@pigment-css/react": "^0.0.2",
 		"@sentry/nextjs": "^7.105.0",
 		"blurhash": "^2.0.5",
@@ -28,7 +29,6 @@
 		"firebase": "10.3.1",
 		"next": "14.2.3",
 		"openai": "^4.47.3",
-		"postgres": "^3.4.4",
 		"react": "^18",
 		"react-dom": "^18",
 		"swr": "^2.2.5"
