+++ src/app/(insert-ad-page)/novel/[chapter]/page.tsx
@@ -1,6 +1,7 @@
 import { styled } from "@pigment-css/react"
 import { Article, VocabularyList, VocabularyProvider } from "./client"
-import { ChapterAPIResponse } from "@/app/novel/types"
+import { Chapter } from "@/app/novel/types"
+import { neon } from "@neondatabase/serverless"
 
 const Container = styled("div")({
 	padding: "20px 20px",
@@ -26,10 +27,11 @@ async function getChapter(chapter: string) {
 		/**
 		 ref: https://stackoverflow.com/questions/76309154/next-js-typeerror-failed-to-parse-url-from-when-targeting-api-route-relati
 		 **/
-		const endpoint = process.env.URL || "http://localhost:3000"
-		const res = await fetch(`${endpoint}/api/novel/chapter?id=${chapter}`)
-		const data: ChapterAPIResponse = await res.json()
-		return data
+
+		const sql = neon(process.env.DATABASE_URL || "")
+		const chapters = (await sql`select * from chapter where id = ${chapter};`) as Chapter[]
+
+		return { chapter: chapters[0] }
 	} catch (e) {
 		throw e
 	}
